requirements.txt
python-dotenv==0.21.1
requests==2.28.2
telegram==0.0.1
schedule~=1.1.0